requirements.txt
torch>=2.0.0
transformers>=4.30.0
assemblyai>=0.17.0
sentencepiece>=0.1.99
pandas>=1.5.0
tqdm>=4.65.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pytubefix>=6.0.0
pydub>=0.25.1